feat: add save trace support

+++ Arbiter.Net/Client/ClientPacket.cs
@@ -1,8 +1,8 @@
 ﻿namespace Arbiter.Net.Client;
 
-public class ClientPacket(byte command, ReadOnlySpan<byte> payload, long? checksum = null) : NetworkPacket(command, payload)
+public class ClientPacket(byte command, ReadOnlySpan<byte> payload, uint? checksum = null) : NetworkPacket(command, payload)
 {
-    public long? Checksum { get; } = checksum;
+    public uint? Checksum { get; } = checksum;
     
     public new ClientCommand Command => Enum.IsDefined(typeof(ClientCommand), base.Command)
         ? (ClientCommand)base.Command
